feat(agents): add header with New Drive/Agent/Session CTAs to Agents list
The Agents console's past-conversations list had no header — it started
immediately below the tab bar with no way to start something new from
inside it. New Session lived only in the sidebar, New Drive only on the
Drives page, and there was no dedicated New Agent affordance anywhere.

Adds AgentsListHeader above the list with all three CTAs:
- New Drive reuses the existing CreateDriveDialog as-is.
- New Session extracts the sidebar's spawn flow (spawn(),
  SpawnSessionPalette, related state) into a shared useSpawnSession hook
  so the header and sidebar don't duplicate the two-step spawn logic.
  AgentsSidebar.tsx is refactored to use the same hook with no behavior
  change.
- New Agent reuses the existing Quick Create palette.

On the global (driveless) Agents page, New Session/New Agent open a new
DrivePickerDialog first, since neither creation path has a drive to act
on without one.
